test(ai): focus tests on stable contracts
