+++ tests/ai/test_step_projection.py
@@ -1,31 +1,23 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-"""Conformance coverage for staged step projection and local streaming."""
+"""Step projection and local stream behavior."""
 
 from datetime import datetime, timezone
 from pathlib import Path
 
 import pytest
 from linktools.ai.agent import AgentBindingSnapshot
-from linktools.ai.agent._output import bind_output
 from linktools.ai.core import ExecutionDeltaType, ExecutionLineageKind, ExecutionStatus
-from linktools.ai.runtime import RuntimeDomain, RuntimeState
-from linktools.ai.runtime._event import ExecutionDelta, LiveExecutionEventBroker
-from linktools.ai.runtime.state import (
-    ExecutionHistoryHeadRecord,
-    ExecutionHistoryState,
-    ExecutionRecord,
+from linktools.ai.runtime import (
+    ExecutionDelta,
+    LiveExecutionEventBroker,
+    RuntimeDomain,
+    RuntimeState,
 )
-from linktools.ai.runtime.state._filesystem import FilesystemStateStore
-from linktools.ai.runtime.state._store import StateTransaction
+from linktools.ai.runtime.state import ExecutionRecord
 from linktools.ai.spec import AgentSpec
 from pydantic_ai.messages import ModelRequest, UserPromptPart
-from pydantic_ai_harness.step_persistence import (
-    ContinuableSnapshot,
-    RunRecord,
-    StepEvent,
-    ToolEffectRecord,
-)
+from pydantic_ai_harness.step_persistence import ContinuableSnapshot, RunRecord, StepEvent
 
 
 def _run() -> RunRecord:
@@ -40,15 +32,14 @@ def _run() -> RunRecord:
 
 
 def _binding_snapshot() -> AgentBindingSnapshot:
-    output = bind_output()
     return AgentBindingSnapshot(
         version=1,
         agent_spec=AgentSpec("agent"),
         model={"route_id": "default", "model_identity": "test:model"},
         selected=(),
         subagents=(),
-        output_mode=output.mode,
-        output_schema=output.schema_definition,
+        output_mode="text",
+        output_schema={"type": "object", "properties": {"text": {"type": "string"}}},
         binding_digest="a" * 64,
     )
 
@@ -85,19 +76,7 @@ async def test_step_events_wait_for_a_safe_snapshot(tmp_path: Path) -> None:
     state = RuntimeState.filesystem(tmp_path / "runtime")
     await state.initialize(namespace="step-io", tenant_id="tenant")
     try:
-        await state.execution.executions.create(_execution())
-        await state.execution.executions.state_store.mutate(
-            lambda transaction: state.execution.executions.insert_history_head_in_transaction(
-                transaction,
-                ExecutionHistoryHeadRecord(
-                    "execution",
-                    "tenant",
-                    ExecutionHistoryState.OPEN,
-                    0,
-                    None,
-                ),
-            )
-        )
+        await state.execution.executions.create_with_history_head(_execution())
         run = _run()
         await state.steps.register_run(run)
         now = datetime.now(timezone.utc)
@@ -115,6 +94,7 @@ async def test_step_events_wait_for_a_safe_snapshot(tmp_path: Path) -> None:
                     agent_name=run.agent_name,
                 )
             )
+
         execution = state.steps.read_store(RuntimeDomain.EXECUTION)
         recovery = state.steps.read_store(RuntimeDomain.RECOVERY)
         assert await execution.list_events(run_id=run.run_id) == []
@@ -136,66 +116,18 @@ async def test_step_events_wait_for_a_safe_snapshot(tmp_path: Path) -> None:
         )
         await state.steps.save_snapshot(snapshot)
         assert await execution.list_events(run_id=run.run_id) == []
+
         await state.steps.flush_execution_projection(
             run.run_id,
             execution_id="execution",
         )
+
         assert len(await execution.list_events(run_id=run.run_id)) == 3
         assert await execution.latest_snapshot(run_id=run.run_id) == snapshot
         assert await recovery.get_run(run_id=run.run_id) == run
         assert await recovery.latest_snapshot(run_id=run.run_id) == snapshot
     finally:
         await state.close()
-
-
-@pytest.mark.asyncio
-async def test_first_recovery_effect_uses_one_filesystem_generation(tmp_path: Path) -> None:
-    state = RuntimeState.filesystem(tmp_path / "runtime")
-    await state.initialize(namespace="effect-io", tenant_id="tenant")
-    try:
-        run = _run()
-        await state.steps.register_run(run)
-        now = datetime.now(timezone.utc)
-        await state.steps.record_tool_effect(
-            ToolEffectRecord(
-                tool_call_id="call",
-                tool_name="tool",
-                run_id=run.run_id,
-                status="started",
-                started_at=now,
-                ended_at=None,
-                idempotency_key="idempotency",
-                effect_summary=None,
-            )
-        )
-        generation = next((path for path in (tmp_path / "runtime" / "recovery").rglob("generation")), None)
-        assert generation is not None
-        assert generation.read_text(encoding="utf-8") == "1"
-    finally:
-        await state.close()
-
-
-@pytest.mark.asyncio
-async def test_filesystem_speculative_callback_runs_once(tmp_path: Path) -> None:
-    store = FilesystemStateStore(
-        tmp_path / "state",
-        namespace="mutation-io",
-        tenant_id="tenant",
-        runtime_domain="execution",
-    )
-    await store.initialize()
-    calls = 0
-
-    async def callback(transaction: StateTransaction) -> int:
-        nonlocal calls
-        calls += 1
-        return await transaction.get_sequence(b"x" * 32)
-
-    try:
-        assert await store.mutate(callback) == 0
-        assert calls == 1
-    finally:
-        await store.close()
 
 
 @pytest.mark.asyncio
@@ -211,6 +143,7 @@ async def test_prepared_local_stream_survives_fast_completion() -> None:
         )
     )
     broker.complete("execution")
+
     subscription = broker.subscribe("execution")
     assert (await subscription.__anext__()).content == "fast"
     with pytest.raises(StopAsyncIteration):
